+++ test/server/registry.spec.lua
@@ -4,7 +4,7 @@ return function()
     describe("simpleLookup", function()
         it("should return 34", function()
             local value = Test:SimpleLookup();
-            expect(value == 34).to.be.ok();
+            expect(value).to.equal(34)
         end)
     end);
 
@@ -12,17 +12,28 @@ return function()
         local value = Test:IntermediateSearch();
 
         it("should return a table", function()
-            expect(type(value) == "table").to.be.ok();
+            expect(type(value)).to.equal("table")
         end)
 
         it("should have a key 'name' equal to 'log'", function()
-            local value = Test:IntermediateSearch();
-            expect(value.name == "log").to.be.ok();
+            expect(value.name).to.equal("log")
         end)
 
         it("should have a key 'quantity' equal to 34", function()
-            expect(value.quantity == 34).to.be.ok();
+            expect(value.quantity).to.equal(34)
         end)
     end);
     
+    describe("advancedSearch", function()
+        local value = Test:AdvancedSearch();
+
+        it("should return a table", function()
+            expect(type(value)).to.equal("table")
+        end)
+
+        it("should have 3 values", function()
+            print(#value == 3)
+            expect(#value).to.equal(3)
+        end)
+    end);
 end