+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Added RegistryModule.remove()
 - Added SearchResult:getFirst()
 
+- Added `exclude` parameter in SearchResult:forEach() callback
+
 ## 1.0.0
 
 - Initial release